+++ src/main/java/com/t1t/t1c/core/GclAgent.java
@@ -1,13 +1,13 @@
-
 package com.t1t.t1c.core;
 
-import java.util.Map;
-import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
+
+import javax.annotation.Generated;
+import java.util.Map;
 
 @Generated("org.jsonschema2pojo")
 public class GclAgent {
@@ -26,24 +26,20 @@ public class GclAgent {
     private Map<String, String> metadata;
     @SerializedName("port")
     @Expose
-    private Integer port;
+    private Long port;
     @SerializedName("username")
     @Expose
     private String username;
 
     /**
-     * 
-     * @return
-     *     The challenge
+     * @return The challenge
      */
     public String getChallenge() {
         return challenge;
     }
 
     /**
-     * 
-     * @param challenge
-     *     The challenge
+     * @param challenge The challenge
      */
     public void setChallenge(String challenge) {
         this.challenge = challenge;
@@ -55,18 +51,14 @@ public class GclAgent {
     }
 
     /**
-     * 
-     * @return
-     *     The hostname
+     * @return The hostname
      */
     public String getHostname() {
         return hostname;
     }
 
     /**
-     * 
-     * @param hostname
-     *     The hostname
+     * @param hostname The hostname
      */
     public void setHostname(String hostname) {
         this.hostname = hostname;
@@ -78,18 +70,14 @@ public class GclAgent {
     }
 
     /**
-     * 
-     * @return
-     *     The lastUpdate
+     * @return The lastUpdate
      */
     public String getLastUpdate() {
         return lastUpdate;
     }
 
     /**
-     * 
-     * @param lastUpdate
-     *     The last_update
+     * @param lastUpdate The last_update
      */
     public void setLastUpdate(String lastUpdate) {
         this.lastUpdate = lastUpdate;
@@ -101,18 +89,14 @@ public class GclAgent {
     }
 
     /**
-     * 
-     * @return
-     *     The metadata
+     * @return The metadata
      */
     public Map<String, String> getMetadata() {
         return metadata;
     }
 
     /**
-     * 
-     * @param metadata
-     *     The metadata
+     * @param metadata The metadata
      */
     public void setMetadata(Map<String, String> metadata) {
         this.metadata = metadata;
@@ -124,41 +108,33 @@ public class GclAgent {
     }
 
     /**
-     * 
-     * @return
-     *     The port
+     * @return The port
      */
-    public Integer getPort() {
+    public Long getPort() {
         return port;
     }
 
     /**
-     * 
-     * @param port
-     *     The port
+     * @param port The port
      */
-    public void setPort(Integer port) {
+    public void setPort(Long port) {
         this.port = port;
     }
 
-    public GclAgent withPort(Integer port) {
+    public GclAgent withPort(Long port) {
         this.port = port;
         return this;
     }
 
     /**
-     * 
-     * @return
-     *     The username
+     * @return The username
      */
     public String getUsername() {
         return username;
     }
 
     /**
-     * 
-     * @param username
-     *     The username
+     * @param username The username
      */
     public void setUsername(String username) {
         this.username = username;
