+++ src/main/java/com/t1t/t1c/ds/DsRegistrationSyncResponse.java
@@ -1,14 +1,14 @@
-
 package com.t1t.t1c.ds;
 
-import java.util.ArrayList;
-import java.util.List;
-import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
+
+import javax.annotation.Generated;
+import java.util.ArrayList;
+import java.util.List;
 
 @Generated("org.jsonschema2pojo")
 public class DsRegistrationSyncResponse {
@@ -27,7 +27,7 @@ public class DsRegistrationSyncResponse {
     private String coreVersion;
     @SerializedName("contextToken")
     @Expose
-    private Integer contextToken;
+    private Long contextToken;
     @SerializedName("containerResponses")
     @Expose
     private List<DsContainerResponse> containerResponses = new ArrayList<DsContainerResponse>();
@@ -36,18 +36,14 @@ public class DsRegistrationSyncResponse {
     private DsAtrList atrList;
 
     /**
-     * 
-     * @return
-     *     The uuid
+     * @return The uuid
      */
     public String getUuid() {
         return uuid;
     }
 
     /**
-     * 
-     * @param uuid
-     *     The uuid
+     * @param uuid The uuid
      */
     public void setUuid(String uuid) {
         this.uuid = uuid;
@@ -59,18 +55,14 @@ public class DsRegistrationSyncResponse {
     }
 
     /**
-     * 
-     * @return
-     *     The activated
+     * @return The activated
      */
     public Boolean getActivated() {
         return activated;
     }
 
     /**
-     * 
-     * @param activated
-     *     The activated
+     * @param activated The activated
      */
     public void setActivated(Boolean activated) {
         this.activated = activated;
@@ -82,18 +74,14 @@ public class DsRegistrationSyncResponse {
     }
 
     /**
-     * 
-     * @return
-     *     The managed
+     * @return The managed
      */
     public Boolean getManaged() {
         return managed;
     }
 
     /**
-     * 
-     * @param managed
-     *     The managed
+     * @param managed The managed
      */
     public void setManaged(Boolean managed) {
         this.managed = managed;
@@ -105,18 +93,14 @@ public class DsRegistrationSyncResponse {
     }
 
     /**
-     * 
-     * @return
-     *     The coreVersion
+     * @return The coreVersion
      */
     public String getCoreVersion() {
         return coreVersion;
     }
 
     /**
-     * 
-     * @param coreVersion
-     *     The coreVersion
+     * @param coreVersion The coreVersion
      */
     public void setCoreVersion(String coreVersion) {
         this.coreVersion = coreVersion;
@@ -128,41 +112,33 @@ public class DsRegistrationSyncResponse {
     }
 
     /**
-     * 
-     * @return
-     *     The contextToken
+     * @return The contextToken
      */
-    public Integer getContextToken() {
+    public Long getContextToken() {
         return contextToken;
     }
 
     /**
-     * 
-     * @param contextToken
-     *     The contextToken
+     * @param contextToken The contextToken
      */
-    public void setContextToken(Integer contextToken) {
+    public void setContextToken(Long contextToken) {
         this.contextToken = contextToken;
     }
 
-    public DsRegistrationSyncResponse withContextToken(Integer contextToken) {
+    public DsRegistrationSyncResponse withContextToken(Long contextToken) {
         this.contextToken = contextToken;
         return this;
     }
 
     /**
-     * 
-     * @return
-     *     The containerResponses
+     * @return The containerResponses
      */
     public List<DsContainerResponse> getContainerResponses() {
         return containerResponses;
     }
 
     /**
-     * 
-     * @param containerResponses
-     *     The containerResponses
+     * @param containerResponses The containerResponses
      */
     public void setContainerResponses(List<DsContainerResponse> containerResponses) {
         this.containerResponses = containerResponses;
@@ -174,18 +150,14 @@ public class DsRegistrationSyncResponse {
     }
 
     /**
-     * 
-     * @return
-     *     The atrList
+     * @return The atrList
      */
     public DsAtrList getAtrList() {
         return atrList;
     }
 
     /**
-     * 
-     * @param atrList
-     *     The atrList
+     * @param atrList The atrList
      */
     public void setAtrList(DsAtrList atrList) {
         this.atrList = atrList;
